alterações realizadas na tela de menus  - adicionado menu sair  -adicionado o bem vindo com e-mail do usuário  - mudança de componente e estilo do botão dos menus  - colocado funcionalidade de signout  - colocado funcionado ao clicar no botão back perguntando ao usuario para sair do sistema
 criado validação para verificar se o usuário está logado e se estiver não solicita para fazer login novamente

 aplicação do botão alterar e campo de id do usuário ao alterar na tela de pessoa

+++ app/src/main/java/com/example/appsenai/entity/Pessoa.java
@@ -4,7 +4,7 @@ import java.io.Serializable;
 import java.util.Random;
 
 public class Pessoa implements Serializable {
-    private long id;
+    private int id;
     private String tipo; // Pode ser A-ALUNO - P-PROFESSOR- G-GERENCIADOR DO SISTEMA
     private String nomeComleto;
     private String email;
@@ -18,7 +18,7 @@ public class Pessoa implements Serializable {
                 "\nE-mail: " + email ;
     }
 
-    public Pessoa(long id, String tipo, String nomeComleto, String email, String senha, String cpf) {
+    public Pessoa(int id, String tipo, String nomeComleto, String email, String senha, String cpf) {
         this.id = id;
         this.tipo = tipo;
         this.nomeComleto = nomeComleto;
@@ -45,11 +45,11 @@ public class Pessoa implements Serializable {
         return false;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
